1. Add Method property to BankEnum for method getListOfCurrenciesRate from all APIBank classes 2. Refactor CashApiRequests 3. Add class CurrencyRate for normalizing 4. Add timeOut to APIBank classes 5. Normalize return of all APIBank classes 6. Change type rateSell, rateBuy, rateCross in MonoBankCurrency to float 7. Rename getListOfCurrenciesRateMonoBank to getListOfCurrenciesRate in MonoBankApi 8. Add condition monoBankCurrency.getCurrencyCodeB() == 980 into filter by currencies

+++ src/main/java/bankApi/currency/MonoBankCurrency.java
@@ -5,9 +5,9 @@ public class MonoBankCurrency {
     private int currencyCodeA;
     private int currencyCodeB;
     private long date;
-    private Double rateSell;
-    private Double rateBuy;
-    private Double rateCross;
+    private float rateSell;
+    private float rateBuy;
+    private float rateCross;
 
     public int getCurrencyCodeA() {
         return currencyCodeA;
@@ -21,15 +21,15 @@ public class MonoBankCurrency {
         return date;
     }
 
-    public Double getRateSell() {
+    public float getRateSell() {
         return rateSell;
     }
 
-    public Double getRateBuy() {
+    public float getRateBuy() {
         return rateBuy;
     }
 
-    public Double getRateCross() {
+    public float getRateCross() {
         return rateCross;
     }
 
@@ -45,15 +45,15 @@ public class MonoBankCurrency {
         this.date = date;
     }
 
-    public void setRateSell(Double rateSell) {
+    public void setRateSell(float rateSell) {
         this.rateSell = rateSell;
     }
 
-    public void setRateBuy(Double rateBuy) {
+    public void setRateBuy(float rateBuy) {
         this.rateBuy = rateBuy;
     }
 
-    public void setRateCross(Double rateCross) {
+    public void setRateCross(float rateCross) {
         this.rateCross = rateCross;
     }
